Extract DPoP logic into separate files
