--- Auth0/DPoP.swift
@@ -1,726 +1,0 @@
-// swiftlint:disable file_length
-import Foundation
-import CryptoKit
-
-// MARK: - Enable/Disable DPoP
-
-public protocol SenderConstraining {
-
-    var dpop: DPoP? { get set }
-
-}
-
-public extension SenderConstraining {
-
-    func useDPoP(keychainTag: String = DPoP.defaultKeychainTag) -> Self {
-        var instance = self
-        instance.dpop = DPoP(keychainTag: keychainTag)
-        return instance
-    }
-
-}
-
-extension SenderConstraining {
-
-    func baseHeaders(accessToken: String, tokenType: String) -> [String: String] {
-        return ["Authorization": "\(tokenType) \(accessToken)"]
-    }
-
-}
-
-// MARK: - Error Type
-
-public struct DPoPError: Auth0Error, Sendable {
-
-    enum Code: Equatable {
-        case secureEnclaveOperationFailed(String)
-        case keychainOperationFailed(String)
-        case cryptoKitOperationFailed(String)
-        case secKeyOperationFailed(String)
-        case other
-        case unknown(String)
-    }
-
-    let code: Code
-
-    init(code: Code, cause: Error? = nil) {
-        self.code = code
-        self.cause = cause
-    }
-
-    /// The underlying `Error` value, if any. Defaults to `nil`.
-    public let cause: Error?
-
-    /// Description of the error.
-    ///
-    /// - Important: You should avoid displaying the error description to the user, it's meant for **debugging** only.
-    public var debugDescription: String {
-        self.appendCause(to: self.message)
-    }
-
-    public static let secureEnclaveOperationFailed: DPoPError = .init(code: .secureEnclaveOperationFailed(""))
-
-    public static let keychainOperationFailed: DPoPError = .init(code: .keychainOperationFailed(""))
-
-    public static let cryptoKitOperationFailed: DPoPError = .init(code: .cryptoKitOperationFailed(""))
-
-    public static let secKeyOperationFailed: DPoPError = .init(code: .secKeyOperationFailed(""))
-
-    /// An unexpected error occurred, and an `Error` value is available.
-    /// The underlying `Error` value can be accessed via the ``Auth0Error/cause-9wuyi`` property.
-    public static let other: DPoPError = .init(code: .other)
-
-    /// An unexpected error occurred, but an `Error` value is not available.
-    /// This error does not include a ``Auth0Error/cause-9wuyi``.
-    public static let unknown: DPoPError = .init(code: .unknown(""))
-
-}
-
-extension DPoPError {
-
-    var message: String {
-        switch self.code {
-        case .secureEnclaveOperationFailed(let message): return message
-        case .cryptoKitOperationFailed(let message): return message
-        case .secKeyOperationFailed(let message): return message
-        case .keychainOperationFailed(let message): return message
-        case .other: return "An unexpected error occurred."
-        case .unknown(let message): return message
-        }
-    }
-
-}
-
-extension DPoPError: Equatable {
-
-    /// Conformance to `Equatable`.
-    public static func == (lhs: DPoPError, rhs: DPoPError) -> Bool {
-        return lhs.code == rhs.code && lhs.localizedDescription == rhs.localizedDescription
-    }
-
-}
-
-public extension DPoPError {
-
-    /// Matches `DPoPError` values in a switch statement.
-    static func ~= (lhs: DPoPError, rhs: DPoPError) -> Bool {
-        return lhs.code == rhs.code
-    }
-
-    /// Matches `Error` values in a switch statement.
-    static func ~= (lhs: DPoPError, rhs: Error) -> Bool {
-        guard let rhs = rhs as? DPoPError else { return false }
-        return lhs.code == rhs.code
-    }
-
-}
-
-// MARK: - DPoP Service
-
-public struct DPoP: Sendable {
-
-    static public let defaultKeychainTag = Bundle.main.bundleIdentifier!
-    static let nonceRequiredErrorCode = "use_dpop_nonce"
-    static var auth0Nonce: String?
-    static private let maxRetries = 1
-
-    private let keyStore: PoPKeyStore
-
-    init(keychainTag: String) {
-        self.keyStore = Self.keyStore(for: keychainTag)
-    }
-
-    static public func addHeaders(to request: inout URLRequest,
-                                  accessToken: String,
-                                  tokenType: String,
-                                  nonce: String? = nil,
-                                  keychainTag: String = defaultKeychainTag) throws(DPoPError) {
-        request.setValue("\(tokenType) \(accessToken)", forHTTPHeaderField: "Authorization")
-
-        guard tokenType.caseInsensitiveCompare("dpop") == .orderedSame else { return }
-        guard let url = request.url, let method = request.httpMethod else {
-            assert(request.url != nil, "The request URL must not be nil.")
-            assert(request.httpMethod != nil, "The request HTTP method must not be nil.")
-            return
-        }
-
-        let proof = try withSerialQueueSync {
-            return try DPoPProof.generate(using: DPoP.keyStore(for: keychainTag),
-                                          url: url,
-                                          method: method,
-                                          nonce: nonce,
-                                          accessToken: accessToken)
-        }
-
-        request.setValue(proof, forHTTPHeaderField: "DPoP")
-    }
-
-    static public func clearKeypair(for keychainTag: String = defaultKeychainTag) throws(DPoPError) {
-        return try withSerialQueueSync {
-            return try Self.keyStore(for: keychainTag).clear()
-        }
-    }
-
-    static public func isNonceRequired(by response: HTTPURLResponse) -> Bool {
-        return DPoPChallenge(from: response)?.errorCode == nonceRequiredErrorCode
-    }
-
-    static func challenge(from response: HTTPURLResponse) -> DPoPChallenge? {
-        return DPoPChallenge(from: response)
-    }
-
-    static func keyStore(for keychainTag: String, useSecureEncave: Bool = SecureEnclave.isAvailable) -> PoPKeyStore {
-        return useSecureEncave ?
-        SecureEnclaveKeyStore(keychainService: keychainTag) :
-        KeychainKeyStore(keychainTag: keychainTag)
-    }
-
-    static func extractNonce(from response: HTTPURLResponse?) -> String? {
-        return response?.value(forHTTPHeaderField: "DPoP-Nonce")
-    }
-
-    static func storeNonce(from response: HTTPURLResponse?) {
-        guard let nonce = extractNonce(from: response) else { return }
-
-        serialQueue.sync {
-            auth0Nonce = nonce
-        }
-    }
-
-    static func shouldRetry(for error: Auth0APIError, retryCount: Int) -> Bool {
-        let isDPoPError = error.code == nonceRequiredErrorCode
-        let isRetryCountExceeded = retryCount >= maxRetries
-
-        return isDPoPError && !isRetryCountExceeded
-    }
-
-    static private func withSerialQueueSync<T>(_ block: () throws -> T) throws(DPoPError) -> T {
-        do {
-            return try serialQueue.sync {
-                return try block()
-            }
-        } catch let error as DPoPError {
-            throw error
-        } catch {
-            throw DPoPError(code: .other, cause: error)
-        }
-    }
-
-    func hasKeypair() throws(DPoPError) -> Bool {
-        return try Self.withSerialQueueSync {
-            return try keyStore.hasPrivateKey()
-        }
-    }
-
-    func shouldGenerateProof(for url: URL, parameters: [String: Any]) throws(DPoPError) -> Bool {
-        if url.lastPathComponent == "token",
-           let grantType = parameters["grant_type"] as? String,
-           grantType != "refresh_token" { return true }
-
-        return try hasKeypair()
-    }
-
-    func generateProof(for request: URLRequest) throws(DPoPError) -> String {
-        let authorizationHeader = request.value(forHTTPHeaderField: "Authorization")
-        let accessToken = authorizationHeader?.components(separatedBy: " ").last
-
-        return try Self.withSerialQueueSync {
-            return try DPoPProof.generate(using: keyStore,
-                                          url: request.url!,
-                                          method: request.httpMethod!,
-                                          nonce: Self.auth0Nonce,
-                                          accessToken: accessToken)
-        }
-    }
-
-    func jkt() throws(DPoPError) -> String {
-        return try Self.withSerialQueueSync {
-            let publicKey = try keyStore.privateKey().publicKey
-            return try ECPublicKeyJWK(publicKey: publicKey).thumbprint()
-        }
-    }
-
-}
-
-// MARK: - Proof Generation
-
-enum DPoPProof {
-
-    static func generate(using keyStore: PoPKeyStore,
-                         url: URL,
-                         method: String,
-                         nonce: String?,
-                         accessToken: String?) throws(DPoPError) -> String {
-        let header = try generateHeader(using: keyStore)
-        let payload = generatePayload(url: url, method: method, nonce: nonce, accessToken: accessToken)
-
-        let headerData: Data
-        let payloadData: Data
-        do {
-            headerData = try JSONSerialization.data(withJSONObject: header, options: [])
-            payloadData = try JSONSerialization.data(withJSONObject: payload, options: [])
-        } catch {
-            throw DPoPError(code: .other, cause: error)
-        }
-
-        let signableParts = "\(headerData.encodeBase64URLSafe()).\(payloadData.encodeBase64URLSafe())"
-        let privateKey = try keyStore.privateKey()
-
-        do {
-            let signature = try privateKey.signature(for: signableParts.data(using: .utf8)!)
-            return "\(signableParts).\(signature.dataRepresentation.encodeBase64URLSafe())"
-        } catch {
-            let message = "Unable to sign the DPoP proof with the private key."
-            throw DPoPError(code: .cryptoKitOperationFailed(message), cause: error)
-        }
-    }
-
-    private static func generateHeader(using keyStore: PoPKeyStore) throws(DPoPError) -> [String: Any] {
-        let publicKey = try keyStore.privateKey().publicKey
-        let jwk = ECPublicKeyJWK(publicKey: publicKey)
-
-        return [
-            "typ": "dpop+jwt",
-            "alg": keyStore.publicKeyJWSIdentifier,
-            "jwk": jwk.toDictionary()
-        ]
-    }
-
-    private static func generatePayload(url: URL,
-                                        method: String,
-                                        nonce: String?,
-                                        accessToken: String?) -> [String: Any] {
-        var urlComponents = URLComponents(url: url, resolvingAgainstBaseURL: true)!
-        urlComponents.query = nil
-        urlComponents.fragment = nil
-
-        var payload: [String: Any] = [
-            "htm": method.uppercased(),
-            "htu": urlComponents.url!.absoluteString,
-            "jti": UUID().uuidString,
-            "iat": Int(Date().timeIntervalSince1970)
-        ]
-        payload["nonce"] = nonce
-
-        if let token = accessToken {
-            let digest = SHA256.hash(data: token.data(using: .utf8)!)
-            payload["ath"] = Data(digest).encodeBase64URLSafe()
-        }
-
-        return payload
-    }
-
-}
-
-// MARK: - Challenge Parsing
-
-struct DPoPChallenge {
-
-    let errorCode: String
-    let errorDescription: String?
-
-    init(errorCode: String, errorDescription: String?) {
-        self.errorCode = errorCode
-        self.errorDescription = errorDescription
-    }
-
-    init?(from response: HTTPURLResponse) {
-        guard let header = response.value(forHTTPHeaderField: "WWW-Authenticate"),
-              header.range(of: "DPoP ", options: .caseInsensitive) != nil else { return nil }
-
-        let valuePattern = #"([\x20-\x21\x23-\x5B\x5D-\x7E]+)"#
-        let errorCodePattern = #"error\s?=\s?""# + valuePattern + "\""
-        let errorDescriptionPattern = #"error_description\s?=\s?""# + valuePattern + "\""
-        let compareOptions: String.CompareOptions = [.caseInsensitive, .regularExpression]
-
-        guard let errorCodeRange = header.range(of: errorCodePattern, options: compareOptions) else {
-            return nil
-        }
-
-        let errorCode = String(header[errorCodeRange])
-
-        if let errorDescriptionRange = header.range(of: errorDescriptionPattern, options: compareOptions) {
-            self.init(errorCode: errorCode, errorDescription: String(header[errorDescriptionRange]))
-        }
-
-        self.init(errorCode: errorCode, errorDescription: nil)
-    }
-
-}
-
-// MARK: - Key Management
-
-struct ECPublicKeyJWK: Encodable {
-
-    enum CodingKeys: String, CodingKey {
-        // swiftlint:disable:next identifier_name
-        case kty, crv, x, y
-    }
-
-    static let jsonEncoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = .sortedKeys
-        return encoder
-    }()
-
-    let kty = "EC"
-    let crv: String
-    let x: String // swiftlint:disable:this identifier_name
-    let y: String // swiftlint:disable:this identifier_name
-
-    init(publicKey: P256.Signing.PublicKey) {
-        self.crv = "P-256"
-        let bytes = publicKey.rawRepresentation
-        self.x = bytes.prefix(32).encodeBase64URLSafe()
-        self.y = bytes.suffix(32).encodeBase64URLSafe()
-    }
-
-    func thumbprint() throws(DPoPError) -> String {
-        let jsonData: Data
-        do {
-            jsonData = try Self.jsonEncoder.encode(self)
-        } catch {
-            throw DPoPError(code: .other, cause: error)
-        }
-
-        return Data(SHA256.hash(data: jsonData)).encodeBase64URLSafe()
-    }
-
-    func toDictionary() -> [String: Any] {
-        return [
-            CodingKeys.kty.rawValue: kty,
-            CodingKeys.crv.rawValue: crv,
-            CodingKeys.x.rawValue: x,
-            CodingKeys.y.rawValue: y
-        ]
-    }
-
-}
-
-// MARK: Secure Enclave Keys
-// From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-
-protocol GenericPasswordConvertible: PoPPrivateKey, CustomStringConvertible {
-
-    /// Creates a key from a raw representation.
-    init<D>(rawRepresentation data: D) throws where D: ContiguousBytes
-
-    /// A raw representation of the key.
-    var rawRepresentation: Data { get }
-
-}
-
-extension GenericPasswordConvertible {
-
-    public var description: String {
-        return self.rawRepresentation.withUnsafeBytes { bytes in
-            return "Key representation contains \(bytes.count) bytes."
-        }
-    }
-
-}
-
-extension SecureEnclave.P256.Signing.PrivateKey: @retroactive CustomStringConvertible {}
-
-extension SecureEnclave.P256.Signing.PrivateKey: GenericPasswordConvertible {
-
-    init<D>(rawRepresentation data: D) throws where D: ContiguousBytes {
-        try self.init(dataRepresentation: data.dataRepresentation)
-    }
-
-    var rawRepresentation: Data {
-        return dataRepresentation  // Contiguous bytes repackaged as a Data instance.
-    }
-
-}
-
-extension ContiguousBytes {
-
-    /// A Data instance created safely from the contiguous bytes without making any copies.
-    var dataRepresentation: Data {
-        return self.withUnsafeBytes { bytes in
-            let cfdata = CFDataCreateWithBytesNoCopy(nil,
-                                                     bytes.baseAddress?.assumingMemoryBound(to: UInt8.self),
-                                                     bytes.count,
-                                                     kCFAllocatorNull)
-            return ((cfdata as NSData?) as Data?) ?? Data(buffer: bytes.assumingMemoryBound(to: UInt8.self))
-        }
-    }
-
-}
-
-// MARK: Non-Secure Enclave Keys
-// From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-
-protocol SecKeyConvertible: PoPPrivateKey, CustomStringConvertible {
-
-    /// Creates a key from an X9.63 representation.
-    init<Bytes>(x963Representation: Bytes) throws where Bytes: ContiguousBytes
-
-    /// An X9.63 representation of the key.
-    var x963Representation: Data { get }
-
-}
-
-extension P256.Signing.PrivateKey: @retroactive CustomStringConvertible {}
-
-// MARK: Private Key Type
-
-protocol PoPPrivateKey: Sendable {
-
-    var publicKey: P256.Signing.PublicKey { get }
-
-    func signature(for data: Data) throws -> P256.Signing.ECDSASignature
-
-}
-
-extension SecureEnclave.P256.Signing.PrivateKey: PoPPrivateKey {
-
-    public var description: String {
-        return self.dataRepresentation.withUnsafeBytes { bytes in
-            return "Key representation contains \(bytes.count) bytes."
-        }
-    }
-
-}
-
-extension P256.Signing.PrivateKey: SecKeyConvertible {
-
-    public var description: String {
-        return self.x963Representation.withUnsafeBytes { bytes in
-            return "Key representation contains \(bytes.count) bytes."
-        }
-    }
-
-}
-
-// MARK: Key Providers
-
-protocol PoPKeyStore: Sendable {
-
-    var publicKeyJWSIdentifier: String { get }
-
-    func hasPrivateKey() throws(DPoPError) -> Bool
-    func privateKey() throws(DPoPError) -> PoPPrivateKey
-    func clear() throws(DPoPError)
-
-}
-
-extension PoPKeyStore {
-
-    var privateKeyIdentifier: String {
-        return "com.auth0.sdk.dpop.privateKey"
-    }
-
-    var publicKeyJWSIdentifier: String {
-        return "ES256"
-    }
-
-}
-
-// Used when Secure Enclave is available
-struct SecureEnclaveKeyStore: PoPKeyStore {
-
-    let keychainService: String
-
-    func hasPrivateKey() throws(DPoPError) -> Bool {
-        if try retrieve(forIdentifier: privateKeyIdentifier) != nil {
-            return true
-        }
-        return false
-    }
-
-    func privateKey() throws(DPoPError) -> PoPPrivateKey {
-        // First, check if the key exists in the keychain
-        if let privateKey = try retrieve(forIdentifier: privateKeyIdentifier) { return privateKey }
-
-        // If not, create a new key and store it in the keychain
-        let privateKey: SecureEnclave.P256.Signing.PrivateKey
-        do {
-            privateKey = try SecureEnclave.P256.Signing.PrivateKey()
-        } catch {
-            let message = "Unable to create a new private key using the Secure Enclave."
-            throw DPoPError(code: .secureEnclaveOperationFailed(message), cause: error)
-        }
-
-        try store(privateKey, forIdentifier: privateKeyIdentifier)
-
-        return privateKey
-    }
-
-    func clear() throws(DPoPError) {
-        let query = baseQuery(forIdentifier: privateKeyIdentifier)
-
-        let status = SecItemDelete(query as CFDictionary)
-        guard (status == errSecSuccess) || (status == errSecItemNotFound) else {
-            let message = "Unable to delete the private key representation from the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-    }
-
-    // From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-    private func store(_ key: GenericPasswordConvertible, forIdentifier identifier: String) throws(DPoPError) {
-        // Describe a generic password
-        var query = baseQuery(forIdentifier: identifier)
-        #if !os(macOS)
-        query[kSecAttrAccessible] = kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly
-        #endif
-        query[kSecValueData] = key.rawRepresentation
-
-        let status = SecItemAdd(query as CFDictionary, nil)
-        guard status == errSecSuccess else {
-            let message = "Unable to store private key representation in the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-    }
-
-    // From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-    private func retrieve(forIdentifier identifier: String) throws(DPoPError) -> GenericPasswordConvertible? {
-        // Seek a generic password with the given account
-        var query = baseQuery(forIdentifier: identifier)
-        query[kSecReturnData] = true
-
-        var item: CFTypeRef?
-        let status = SecItemCopyMatching(query as CFDictionary, &item)
-
-        guard status != errSecItemNotFound else { return nil }
-        guard status == errSecSuccess else {
-            let message = "Unable to retrieve the private key representation from the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-        guard let data = item as? Data else {
-            throw DPoPError(code: .unknown("Unable to cast the retrieved private key representation to Data."))
-        }
-
-        do {
-            return try SecureEnclave.P256.Signing.PrivateKey(rawRepresentation: data)
-        } catch {
-            let message = "Unable to recreate a Secure Enclave private key from the retrieved data."
-            throw DPoPError(code: .secureEnclaveOperationFailed(message), cause: error)
-        }
-    }
-
-    private func baseQuery(forIdentifier identifier: String) -> [CFString: Any] {
-        return [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrService: keychainService,
-            kSecAttrAccount: identifier
-        ]
-    }
-
-}
-
-// Used when Secure Enclave is not available
-struct KeychainKeyStore: PoPKeyStore {
-
-    let keychainTag: Data
-
-    init(keychainTag: String) {
-        self.keychainTag = keychainTag.data(using: .utf8)!
-    }
-
-    func hasPrivateKey() throws(DPoPError) -> Bool {
-        return try retrieve(forIdentifier: privateKeyIdentifier) != nil
-    }
-
-    // When SecureEnclave is not available, we use a simple keychain store
-    func privateKey() throws(DPoPError) -> PoPPrivateKey {
-        // First, check if the key exists in the keychain
-        if let privateKey = try retrieve(forIdentifier: privateKeyIdentifier) { return privateKey }
-
-        // If not, create a new key and store it in the keychain
-        let privateKey = P256.Signing.PrivateKey()
-        try store(privateKey, forIdentifier: privateKeyIdentifier)
-
-        return privateKey
-    }
-
-    func clear() throws(DPoPError) {
-        let query = baseQuery(forIdentifier: privateKeyIdentifier)
-
-        let status = SecItemDelete(query as CFDictionary)
-        guard (status == errSecSuccess) || (status == errSecItemNotFound) else {
-            let message = "Unable to delete the private key from the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-    }
-
-    // From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-    private func retrieve(forIdentifier identifier: String) throws(DPoPError) -> SecKeyConvertible? {
-        // Seek an elliptic-curve key with a given identifier
-        var query = baseQuery(forIdentifier: privateKeyIdentifier)
-        query[kSecAttrKeyType] = kSecAttrKeyTypeECSECPrimeRandom
-        query[kSecReturnRef] = true
-
-        // Find and cast the result as a SecKey instance
-        var item: CFTypeRef?
-        let status = SecItemCopyMatching(query as CFDictionary, &item)
-
-        guard status != errSecItemNotFound else { return nil }
-        guard status == errSecSuccess else {
-            let message = "Unable to retrieve the private key from the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-        // Ensure the item is a SecKey so the force cast is safe
-        guard CFGetTypeID(item) == SecKeyGetTypeID() else {
-            throw DPoPError(code: .unknown("The item retrieved from the Keychain is not a SecKey."))
-        }
-
-        // swiftlint:disable:next force_cast
-        let secKey = item as! SecKey
-
-        // Convert the SecKey into a CryptoKit key
-        var error: Unmanaged<CFError>?
-        guard let data = SecKeyCopyExternalRepresentation(secKey, &error) as Data?, error == nil else {
-            let message = "Unable to copy the external representation of the retrieved SecKey."
-            throw DPoPError(code: .secKeyOperationFailed(message), cause: error?.takeUnretainedValue())
-        }
-
-        do {
-            return try P256.Signing.PrivateKey(x963Representation: data)
-        } catch {
-            let message = "Unable to create a CryptoKit private key from the retrieved SecKey."
-            throw DPoPError(code: .cryptoKitOperationFailed(message), cause: error)
-        }
-    }
-
-    // From Apple's sample code at https://developer.apple.com/documentation/cryptokit/storing_cryptokit_keys_in_the_keychain
-    private func store(_ key: SecKeyConvertible, forIdentifier identifier: String) throws(DPoPError) {
-        // Describe the key
-        let attributes = [
-            kSecAttrKeyType: kSecAttrKeyTypeECSECPrimeRandom,
-            kSecAttrKeyClass: kSecAttrKeyClassPrivate
-        ] as [String: Any]
-
-        // Get a SecKey representation
-        var error: Unmanaged<CFError>?
-        guard let secKey = SecKeyCreateWithData(key.x963Representation as CFData,
-                                                attributes as CFDictionary,
-                                                &error), error == nil else {
-            let message = "Unable to create a SecKey representation from the private key."
-            throw DPoPError(code: .secKeyOperationFailed(message), cause: error?.takeUnretainedValue())
-        }
-
-        // Describe the add operation
-        var query = baseQuery(forIdentifier: privateKeyIdentifier)
-        query[kSecAttrAccessible] = kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly
-        query[kSecValueRef] = secKey
-
-        // Add the key to the keychain
-        let status = SecItemAdd(query as CFDictionary, nil)
-        guard status == errSecSuccess else {
-            let message = "Unable to store the private key in the Keychain. OSStatus: \(status)."
-            throw DPoPError(code: .keychainOperationFailed(message))
-        }
-    }
-
-    private func baseQuery(forIdentifier identifier: String) -> [CFString: Any] {
-        return [
-            kSecClass: kSecClassKey,
-            kSecAttrApplicationTag: keychainTag,
-            kSecAttrApplicationLabel: identifier,
-            kSecUseDataProtectionKeychain: true
-        ]
-    }
-
-}
