+++ Auth0/DPoP/DPoPKeyStore.swift
@@ -1,0 +1,121 @@
+import Foundation
+import CryptoKit
+
+// MARK: Key Providers
+
+protocol DPoPKeyStore: Sendable {
+
+    var publicKeyJWSIdentifier: String { get }
+
+    func hasPrivateKey() throws(DPoPError) -> Bool
+    func privateKey() throws(DPoPError) -> PoPPrivateKey
+    func clear() throws(DPoPError)
+
+}
+
+extension DPoPKeyStore {
+
+    var privateKeyIdentifier: String {
+        return "com.auth0.sdk.dpop.privateKey"
+    }
+
+    var publicKeyJWSIdentifier: String {
+        return "ES256"
+    }
+
+}
+
+// MARK: Private Key Type
+
+protocol PoPPrivateKey: Sendable {
+
+    var publicKey: P256.Signing.PublicKey { get }
+
+    func signature(for data: Data) throws -> P256.Signing.ECDSASignature
+
+}
+
+extension SecureEnclave.P256.Signing.PrivateKey: PoPPrivateKey {
+
+    public var description: String {
+        return self.dataRepresentation.withUnsafeBytes { bytes in
+            return "Key representation contains \(bytes.count) bytes."
+        }
+    }
+
+}
+
+extension P256.Signing.PrivateKey: SecKeyConvertible {
+
+    public var description: String {
+        return self.x963Representation.withUnsafeBytes { bytes in
+            return "Key representation contains \(bytes.count) bytes."
+        }
+    }
+
+}
+
+// MARK: Secure Enclave Keys
+
+protocol GenericPasswordConvertible: PoPPrivateKey, CustomStringConvertible {
+
+    /// Creates a key from a raw representation.
+    init<D>(rawRepresentation data: D) throws where D: ContiguousBytes
+
+    /// A raw representation of the key.
+    var rawRepresentation: Data { get }
+
+}
+
+extension GenericPasswordConvertible {
+
+    public var description: String {
+        return self.rawRepresentation.withUnsafeBytes { bytes in
+            return "Key representation contains \(bytes.count) bytes."
+        }
+    }
+
+}
+
+extension SecureEnclave.P256.Signing.PrivateKey: @retroactive CustomStringConvertible {}
+
+extension SecureEnclave.P256.Signing.PrivateKey: GenericPasswordConvertible {
+
+    init<D>(rawRepresentation data: D) throws where D: ContiguousBytes {
+        try self.init(dataRepresentation: data.dataRepresentation)
+    }
+
+    var rawRepresentation: Data {
+        return dataRepresentation  // Contiguous bytes repackaged as a Data instance.
+    }
+
+}
+
+extension ContiguousBytes {
+
+    /// A Data instance created safely from the contiguous bytes without making any copies.
+    var dataRepresentation: Data {
+        return self.withUnsafeBytes { bytes in
+            let cfdata = CFDataCreateWithBytesNoCopy(nil,
+                                                     bytes.baseAddress?.assumingMemoryBound(to: UInt8.self),
+                                                     bytes.count,
+                                                     kCFAllocatorNull)
+            return ((cfdata as NSData?) as Data?) ?? Data(buffer: bytes.assumingMemoryBound(to: UInt8.self))
+        }
+    }
+
+}
+
+// MARK: Non-Secure Enclave Keys
+
+protocol SecKeyConvertible: PoPPrivateKey, CustomStringConvertible {
+
+    /// Creates a key from an X9.63 representation.
+    init<Bytes>(x963Representation: Bytes) throws where Bytes: ContiguousBytes
+
+    /// An X9.63 representation of the key.
+    var x963Representation: Data { get }
+
+}
+
+extension P256.Signing.PrivateKey: @retroactive CustomStringConvertible {}
